backend/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pydantic==2.5.2
python-dateutil==2.8.2